set up models and migrated

+++ user/models.py
@@ -3,8 +3,8 @@ from django.contrib.auth.models import AbstractUser
 # Create your models here.
 
 class User(AbstractUser):
-    lifetimeWeightLifted = models.IntegerField(default=0)
-    
+    lifetimeWeightLifted = models.DecimalField(default=0, max_digits=12, decimal_places=1)
+
